ossystems-srcrev-handler.bbclass: Import class
This class has been used by several customers from O.S. Systems.

It helps the management of several project components of a product.
The class uses a file to describe the recipe and the version to use
for stable releases but also easy the use of top of tree builds of
those components for development purposes.

Change-Id: I70855c52b4793093d7fee03992e2e2e99a9e5bfb
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>