systemd: Reduce watchdog pings in timeout interval
The watchdog ping is performed for every iteration of manager event
loop. This results in a lot of ioctls on watchdog device driver
especially during boot or if services are aggressively using sd_notify.
Depending on the watchdog device driver this may have performance impact
on embedded systems. The patch skips sending the watchdog to device
driver if the ping is requested before half of the watchdog timeout.

Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>