zram-init: Fix Source Uri and update License
This commit fixes the SRC_URI for the zram-init recipe. It sets the scheme to
git:// to ensure a Git clone is done and uses protocol=https, the current
standard protocol. This change ensures the zram-init recipe can fetch the
source code using the supported HTTPS protocol. The change in the LICENSE field
fixes the obsolete-license warning.

Signed-off-by: Flarno <flarno11@gmail.com>