telegraf: Fix Source Uri and allow network access
This commit fixes the SRC_URI for the telegraf recipe. It sets the scheme to
git:// to ensure a Git clone is done and uses protocol=https, the current
standard protocol. This change ensures the telegraf recipe can fetch the source
code using the supported HTTPS protocol and access the internet during the
compilation.

Signed-off-by: Flarno <flarno11@gmail.com>